抽离failureManager
